Version 10.4.11-ii

+++ Src/SLib/sdate.cpp
@@ -1446,14 +1446,14 @@ int LDATE::getclass() const
 
 int LDATE::encode(int d, int m, int y)
 {
-	int    shift;
+	int16  shift; // @v10.4.12 @fix int-->int16
 	int    x;
 	int    d_ = 0, m_ = 0, y_ = 0;
 	if(d == ANY_DATE_VALUE) {
 		d_ = ANY_DAYITEM_VALUE;
 	}
 	else if(d & REL_DATE_MASK) {
-		shift = static_cast<int>(LoWord(d));
+		shift = static_cast<int16>(LoWord(d));
 		if(m == -1 && y == -1) {
 			v = MakeLong(shift, 0x8000);
 			return 1;
@@ -1469,7 +1469,7 @@ int LDATE::encode(int d, int m, int y)
 		}
 	}
 	else if(d & THRSMDAY_DATE_MASK) {
-		shift = static_cast<int>(LoWord(d));
+		shift = static_cast<int16>(LoWord(d));
 		if(shift >= 1 && shift <= 31) {
 			v = MakeLong(shift, 0x2000);
 			return 1;
@@ -1485,7 +1485,7 @@ int LDATE::encode(int d, int m, int y)
 		m_ = ANY_MONITEM_VALUE;
 	}
 	else if(m & REL_DATE_MASK) {
-		shift = static_cast<int>(LoWord(m));
+		shift = static_cast<int16>(LoWord(m));
 		if(shift < 0)
 			x = (shift <= -24) ? (0x40 | 24) : (0x40 | (-shift));
 		else if(shift > 0)
@@ -1500,7 +1500,7 @@ int LDATE::encode(int d, int m, int y)
 		y_ = ANY_YEARITEM_VALUE;
 	}
 	else if(y & REL_DATE_MASK) {
-		shift = static_cast<int>(LoWord(y));
+		shift = static_cast<int16>(LoWord(y));
 		if(shift < 0)
 			x = (shift <= -255) ? (0x0400 | 255) : (0x0400 | (-shift));
 		else if(shift > 0)
